Get NFL in-game win probabilities for a specific game from ESPN

Get NFL in-game win probabilities for a specific game from ESPN

get_espn_win_prob(game_id)

Arguments

game_id

Character string - can be acquired from the website of a specific game, or from espnscrapeR::get_nfl_schedule()

Value

Returns a tibble

Examples

# Get win prob from specific game
get_espn_win_prob(game_id = "401030956")
#> # A tibble: 160 × 9
#>    row_id home_team_id away_te…¹ tie_p…² home_…³ away_…⁴ seque…⁵ play_id game_id
#>     <int> <chr>        <chr>       <dbl>   <dbl>   <dbl> <chr>   <chr>   <chr>  
#>  1      1 23           17              0   0.599   0.401 100     401030… 401030…
#>  2      2 23           17              0   0.605   0.395 3600    401030… 401030…
#>  3      3 23           17              0   0.597   0.403 5100    401030… 401030…
#>  4      4 23           17              0   0.592   0.408 7700    401030… 401030…
#>  5      5 23           17              0   0.62    0.38  10200   401030… 401030…
#>  6      6 23           17              0   0.619   0.381 12700   401030… 401030…
#>  7      7 23           17              0   0.637   0.363 14900   401030… 401030…
#>  8      8 23           17              0   0.656   0.344 17100   401030… 401030…
#>  9      9 23           17              0   0.647   0.353 20200   401030… 401030…
#> 10     10 23           17              0   0.683   0.317 22400   401030… 401030…
#> # … with 150 more rows, and abbreviated variable names ¹​away_team_id,
#> #   ²​tie_percentage, ³​home_win_percentage, ⁴​away_win_percentage,
#> #   ⁵​sequence_number