Get NFL boxscore for players for a specific game from ESPN's API

Get NFL boxscore for players for a specific game from ESPN's API

get_nfl_boxscore_players(game_id)

Arguments

game_id
character

Value

Returns a tibble

Examples

# Get NFL play-by-play for a specific game
get_nfl_boxscore_players(game_id = "300912027")

#> # A tibble: 60 x 83
#>    game_id game_uid season season_type date   week home_away winner team_id
#>    <chr>   <chr>     <int>       <int> <chr> <int> <chr>      <int> <chr>  
#>  1 300912… s:20~l:…   2010           2 2010…     1 away           0 5      
#>  2 300912… s:20~l:…   2010           2 2010…     1 away           0 5      
#>  3 300912… s:20~l:…   2010           2 2010…     1 away           0 5      
#>  4 300912… s:20~l:…   2010           2 2010…     1 away           0 5      
#>  5 300912… s:20~l:…   2010           2 2010…     1 away           0 5      
#>  6 300912… s:20~l:…   2010           2 2010…     1 away           0 5      
#>  7 300912… s:20~l:…   2010           2 2010…     1 away           0 5      
#>  8 300912… s:20~l:…   2010           2 2010…     1 away           0 5      
#>  9 300912… s:20~l:…   2010           2 2010…     1 away           0 5      
#> 10 300912… s:20~l:…   2010           2 2010…     1 away           0 5      
#> # … with 50 more rows, and 74 more variables: team_uid <chr>, team_city <chr>,
#> #   team_name <chr>, team_abb <chr>, team_full <chr>, team_color <chr>,
#> #   team_color_alt <chr>, team_logo <chr>, team_order <int>, team_score <int>,
#> #   team_record <chr>, player_id <chr>, player_uid <chr>, player_guid <chr>,
#> #   first_name <chr>, last_name <chr>, full_name <chr>, passing_att <int>,
#> #   passing_cmp <int>, passing_yds <dbl>, passing_avg <dbl>, passing_td <dbl>,
#> #   passing_int <dbl>, passing_sacks <dbl>, passing_sack_yards <dbl>,
#> #   passing_qbr <dbl>, passing_rtg <dbl>, rushing_car <dbl>, rushing_yds <dbl>,
#> #   rushing_avg <dbl>, rushing_td <dbl>, rushing_long <dbl>,
#> #   receiving_rec <dbl>, receiving_yds <dbl>, receiving_avg <dbl>,
#> #   receiving_td <dbl>, receiving_long <dbl>, receiving_tgts <dbl>,
#> #   fumbles_fum <dbl>, fumbles_lost <dbl>, fumbles_rec <dbl>,
#> #   defensive_tot <dbl>, defensive_solo <dbl>, defensive_sacks <dbl>,
#> #   defensive_tfl <dbl>, defensive_pd <dbl>, defensive_qb_hts <dbl>,
#> #   defensive_td <dbl>, interceptions_int <dbl>, interceptions_yds <dbl>,
#> #   interceptions_td <dbl>, kickreturns_no <dbl>, kickreturns_yds <dbl>,
#> #   kickreturns_avg <dbl>, kickreturns_long <dbl>, kickreturns_td <dbl>,
#> #   puntreturns_no <dbl>, puntreturns_yds <dbl>, puntreturns_avg <dbl>,
#> #   puntreturns_long <dbl>, puntreturns_td <dbl>, kicking_fg_att <dbl>,
#> #   kicking_fg_made <dbl>, kicking_pct <dbl>, kicking_long <dbl>,
#> #   kicking_xp_att <dbl>, kicking_xp_made <dbl>, kicking_pts <dbl>,
#> #   punting_no <dbl>, punting_yds <dbl>, punting_avg <dbl>, punting_tb <dbl>,
#> #   punting_in_20 <dbl>, punting_long <dbl>