Get NFL in-game win probabilities for a specific game from ESPN

Get NFL in-game win probabilities for a specific game from ESPN

get_espn_win_prob(game_id)

Arguments

game_id
Character string - can be acquired from the website of a specific game, or from espnscrapeR::get_nfl_schedule()

Value

Returns a tibble

Examples

# Get win prob from specific game
get_espn_win_prob(game_id = "401030956")

#> # A tibble: 160 x 9
#>    row_id home_team_id away_team_id tie_percentage home_win_percentage
#>     <int> <chr>        <chr>                 <dbl>               <dbl>
#>  1      1 23           17                        0               0.599
#>  2      2 23           17                        0               0.605
#>  3      3 23           17                        0               0.597
#>  4      4 23           17                        0               0.592
#>  5      5 23           17                        0               0.62 
#>  6      6 23           17                        0               0.619
#>  7      7 23           17                        0               0.637
#>  8      8 23           17                        0               0.656
#>  9      9 23           17                        0               0.647
#> 10     10 23           17                        0               0.683
#> # … with 150 more rows, and 4 more variables: away_win_percentage <dbl>,
#> #   sequence_number <chr>, play_id <chr>, game_id <chr>