Get some of Lee Sharpe's datasets from GitHub

Get some of Lee Sharpe's datasets from GitHub

get_sharpe_data(dataset = "games")

Arguments

dataset
Character only, one of "airports", "closing_lines", "draft_picks", "draft_values", "games", "logos", "pff_pfr_map_v1", "positions", "rosters", "sc_lines", "standings", "teamcolors", "teams", "trades", "win_totals"

Value

tibble

Examples

# Get NFL games and their outcomes
get_sharpe_data(dataset = "games")

#> Rows: 6124 Columns: 41

#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr  (17): game_id, game_type, weekday, away_team, home_team, location, roof...
#> dbl  (22): season, week, away_score, home_score, result, total, overtime, ol...
#> date  (1): gameday
#> time  (1): gametime

#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

#> # A tibble: 6,124 × 41
#>    game_id         season game_type  week gameday    weekday gametime away_team
#>    <chr>            <dbl> <chr>     <dbl> <date>     <chr>   <time>   <chr>    
#>  1 1999_01_MIN_ATL   1999 REG           1 1999-09-12 Sunday     NA    MIN      
#>  2 1999_01_KC_CHI    1999 REG           1 1999-09-12 Sunday     NA    KC       
#>  3 1999_01_PIT_CLE   1999 REG           1 1999-09-12 Sunday     NA    PIT      
#>  4 1999_01_OAK_GB    1999 REG           1 1999-09-12 Sunday     NA    OAK      
#>  5 1999_01_BUF_IND   1999 REG           1 1999-09-12 Sunday     NA    BUF      
#>  6 1999_01_SF_JAX    1999 REG           1 1999-09-12 Sunday     NA    SF       
#>  7 1999_01_CAR_NO    1999 REG           1 1999-09-12 Sunday     NA    CAR      
#>  8 1999_01_NE_NYJ    1999 REG           1 1999-09-12 Sunday     NA    NE       
#>  9 1999_01_ARI_PHI   1999 REG           1 1999-09-12 Sunday     NA    ARI      
#> 10 1999_01_DET_SEA   1999 REG           1 1999-09-12 Sunday     NA    DET      
#> # … with 6,114 more rows, and 33 more variables: away_score <dbl>,
#> #   home_team <chr>, home_score <dbl>, location <chr>, result <dbl>,
#> #   total <dbl>, overtime <dbl>, old_game_id <dbl>, espn <dbl>,
#> #   away_rest <dbl>, home_rest <dbl>, away_moneyline <dbl>,
#> #   home_moneyline <dbl>, spread_line <dbl>, away_spread_odds <dbl>,
#> #   home_spread_odds <dbl>, total_line <dbl>, under_odds <dbl>,
#> #   over_odds <dbl>, div_game <dbl>, roof <chr>, surface <chr>, temp <dbl>, …

# Get team abbreviations/names/etc by year
get_sharpe_data(dataset = "teams")

#> Rows: 640 Columns: 18

#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (14): team, nfl, nfl_team_id, espn, pfr, pfflabel, fo, full, location, s...
#> dbl  (4): season, pff, sbr, sbr_wins

#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

#> # A tibble: 640 × 18
#>    season team  nfl   nfl_team_id espn  pfr     pff pfflabel fo    full         
#>     <dbl> <chr> <chr> <chr>       <chr> <chr> <dbl> <chr>    <chr> <chr>        
#>  1   2002 ARI   ARI   3800        ARI   CRD      NA NA       ARI   Arizona Card…
#>  2   2002 ATL   ATL   0200        ATL   ATL      NA NA       ATL   Atlanta Falc…
#>  3   2002 BAL   BAL   0325        BAL   RAV      NA NA       BAL   Baltimore Ra…
#>  4   2002 BUF   BUF   0610        BUF   BUF      NA NA       BUF   Buffalo Bills
#>  5   2002 CAR   CAR   0750        CAR   CAR      NA NA       CAR   Carolina Pan…
#>  6   2002 CHI   CHI   0810        CHI   CHI      NA NA       CHI   Chicago Bears
#>  7   2002 CIN   CIN   0920        CIN   CIN      NA NA       CIN   Cincinnati B…
#>  8   2002 CLE   CLE   1050        CLE   CLE      NA NA       CLE   Cleveland Br…
#>  9   2002 DAL   DAL   1200        DAL   DAL      NA NA       DAL   Dallas Cowbo…
#> 10   2002 DEN   DEN   1400        DEN   DEN      NA NA       DEN   Denver Bronc…
#> # … with 630 more rows, and 8 more variables: location <chr>,
#> #   short_location <chr>, nickname <chr>, hyphenated <chr>, sbr <dbl>,
#> #   sbr_wins <dbl>, sbr_name <chr>, draft_kings <chr>